requirements.txt
flask==1.1.2
werkzeug==2.0.1
itsdangerous==2.0.1
click==8.0.1
colorama==0.4.4
Jinja2==2.11.3
MarkupSafe==1.1.1
google-cloud-storage==1.41.0